phase2_mqtt/requirements.txt
paho-mqtt>=2.0.0
numpy>=2.1.0
scikit-learn>=1.5.2
joblib
